Sample page for the graphs website will render.
This html file renders the core lower-division courses, in addition to ICS 139W and CS 161.
We need to figure out how to render graphs that include particular specializations.

We have to work to make co-requisite courses into a single node. 
Ex) make ICS 53 and ICS 53-L a single node.


We also have to figure out how to handle courses that have "and"s and "or"s in them.
Ex.) I&C SCI 6D and (MATH 6G or MATH 3A or I&C SCI 6N) and MATH 2B and (I&C SCI 23 or CSE 23 or I&C SCI H23 or I&C SCI 46 or CSE 46) and (COMPSCI 112 or COMPSCI 116 or COMPSCI 171 or COMPSCI 178).