Sample schedule JSON file used to cross reference
This is a JSON file that contains all of the courses in the sample schedule for CS, graciously provided by Jansen.

You're going to have to change the collect_course.py module to load this file into the "data" variable. I kept it in my downloads, but you should probably move it to somewhere else.



Jansen parsed this using Parsehub, which seems like it'd be a very useful tool to scrape the specializations and emphases from the UCI course catalog.