Added even ~more~ graphs that can be rendered by the dropdown
I've added more graphs, up until "networked systems" in the computer science specializations.


An issue that I wasn't even considering is how to tackle the problem of "select n courses from the following list".

Do we want to show all of the nodes? This implies that all of the nodes have to be taken, which isn't the case.
Do we color code them and add a legend? This adds a layer of complexity that we may not want to deal with.
Do we want to have a bubble node that will render a particular path, adding nodes and displaying them if necessary?

I'm really not sure how to proceed.


An idea I have is allow the user to select their particular schedule from a set of all possible nodes.